requirements.txt
flask==2.3.3
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
diffusers==0.24.0
accelerate==0.24.1 